+++ pytest_asyncio_network_simulator/utils.py
@@ -1,5 +1,10 @@
 import asyncio
-from typing import Tuple
+from typing import (
+    Any,
+    Callable,
+    Tuple,
+)
 
 
+ConnectionCallback = Callable[[asyncio.StreamReader, asyncio.StreamWriter], Any]
 ReaderWriterPair = Tuple[asyncio.StreamReader, asyncio.StreamWriter]
